Added nonworking doSingleUpkeep
Need to figure out a safe way to approve the tokens that should be compounded without calling the ERC-20 interface compliant "approve" method. It is dangerous because a instance of the ERC-20 standard contract can override the standard and call any code within the scope of the Compounder contract. Perhaps an intermediary contract could be used.