Created the nonworking doSingleUpKeep
doSingleUpKeep currently does not work because the tokens needed in IncreaseLiquidityParams need to be approved. However, I recognize that "approving" any old ERC-20 contract is dangerous behavior because the ERC-20 implementation of "approve" can be overriden with malicious code, and when called from the scope of the Compound contract, can be extremely dangerous. possibly an intermediary contract to handle the upkeeping per pool can work?